+++ test/utils/server.js
@@ -1,6 +1,5 @@
 var https = require('https');
 var express = require('express');
-var pem = require('pem');
 var rimraf = require('rimraf');
 var nineTrack = require('../../');
 
@@ -50,43 +49,13 @@ exports.run = function (port, middlewares) {
   }, port, middlewares);
 };
 
-var x509, certsCallback;
-
-exports.certs = function (cb) {
-  if (typeof x509 !== 'undefined') {
-    cb(x509);
-  } else {
-    certsCallback = cb;
-  }
-};
-
-exports.runHttps = function (port, middlewares, options) {
-  // Generate an HTTPS certificate
-  before(function generateCertificate (done) {
-    pem.createCertificate({days: 1, selfSigned: true}, function saveCertificate (err, keys) {
-      this.certificate = keys;
-      x509 = {key: keys.clientKey, cert: keys.certificate};
-      pem.getPublicKey(x509.cert, function savePublicCert (publicCert) {
-        x509.cert = publicCert;
-        if (typeof certsCallback === 'function') {
-          certsCallback(x509);
-        }
-      });
-      done(err);
-    });
-  });
-  after(function cleanupCertificate () {
-    delete this.certificate;
-  });
-
-  // Enable our server to require client certs, use defaults otherwise
-  options = options || {requestCert: false, rejectUnauthorized: false};
-
+exports.runHttps = function (port, certs, options, middlewares) {
   // Start the HTTPS server with said certificate
   exports._run(function startHttpsServer (app, port) {
     var server = https.createServer({
-      key: this.certificate.serviceKey,
-      cert: this.certificate.certificate,
+      key: certs.serverKey,
+      cert: certs.serverCert,
+      ca: certs.ca,
       requestCert: options.requestCert,
       rejectUnauthorized: options.rejectUnauthorized
     }, app);
